refactor: api streaming response handling
